chore: move hyperlane_db from `base` to `core`
