+++ rust/agents/relayer/src/merkle_tree/processor.rs
@@ -7,7 +7,7 @@ use std::{
 use async_trait::async_trait;
 use derive_new::new;
 use eyre::Result;
-use hyperlane_base::db::HyperlaneRocksDB;
+use hyperlane_core::db::HyperlaneRocksDB;
 use hyperlane_core::{HyperlaneDomain, MerkleTreeInsertion};
 use prometheus::IntGauge;
 use tokio::sync::RwLock;
