+++ rust/agents/relayer/src/relayer.rs
@@ -9,12 +9,12 @@ use derive_more::AsRef;
 use eyre::Result;
 use futures_util::future::try_join_all;
 use hyperlane_base::{
-    db::{HyperlaneRocksDB, DB},
     metrics::{AgentMetrics, MetricsUpdater},
     settings::ChainConf,
     BaseAgent, ChainMetrics, ContractSyncMetrics, ContractSyncer, CoreMetrics, HyperlaneAgentCore,
     SyncOptions,
 };
+use hyperlane_core::db::{HyperlaneRocksDB, DB};
 use hyperlane_core::{
     HyperlaneDomain, HyperlaneMessage, InterchainGasPayment, MerkleTreeInsertion, QueueOperation,
     H512, U256,
@@ -538,7 +538,6 @@ impl Relayer {
         let serial_submitter = SerialSubmitter::new(
             destination.clone(),
             receiver,
-            Arc::new(self.dbs),
             retry_receiver_channel,
             SerialSubmitterMetrics::new(&self.core.metrics, destination),
             batch_size,
