+++ rust/agents/validator/src/validator.rs
@@ -10,14 +10,13 @@ use tokio::{task::JoinHandle, time::sleep};
 use tracing::{error, info, info_span, instrument::Instrumented, warn, Instrument};
 
 use hyperlane_base::{
-    db::{HyperlaneRocksDB, DB},
-    metrics::AgentMetrics,
-    settings::ChainConf,
-    BaseAgent, ChainMetrics, CheckpointSyncer, ContractSyncMetrics, ContractSyncer, CoreMetrics,
-    HyperlaneAgentCore, MetricsUpdater, SequencedDataContractSync,
+    metrics::AgentMetrics, settings::ChainConf, BaseAgent, ChainMetrics, CheckpointSyncer,
+    ContractSyncMetrics, ContractSyncer, CoreMetrics, HyperlaneAgentCore, MetricsUpdater,
+    SequencedDataContractSync,
 };
 
 use hyperlane_core::{
+    db::{HyperlaneRocksDB, DB},
     Announcement, ChainResult, HyperlaneChain, HyperlaneContract, HyperlaneDomain, HyperlaneSigner,
     HyperlaneSignerExt, Mailbox, MerkleTreeHook, MerkleTreeInsertion, TxOutcome, ValidatorAnnounce,
     H256, U256,
