--- rust/hyperlane-base/src/db/mod.rs
@@ -1,2 +1,0 @@
-pub use rocks::*;
-mod rocks;
