+++ rust/hyperlane-base/src/lib.rs
@@ -28,8 +28,5 @@ pub use traits::*;
 mod types;
 pub use types::*;
 
-/// Hyperlane database utils
-pub mod db;
-
 #[cfg(feature = "oneline-eyre")]
 pub mod oneline_eyre;
