+++ rust/hyperlane-core/Cargo.toml
@@ -27,17 +27,21 @@ ethers-providers = { workspace = true, optional = true }
 eyre.workspace = true
 fixed-hash.workspace = true
 futures = { workspace = true, optional = true }
+futures-util.workspace = true
 getrandom.workspace = true
 hex.workspace = true
+paste.workspace = true
 itertools.workspace = true
 num = { workspace = true, features = ["serde"] }
 num-derive.workspace = true
 num-traits.workspace = true
+rocksdb.workspace = true
 serde = { workspace = true }
 serde_json = { workspace = true }
 sha3 = { workspace = true }
 strum = { workspace = true, optional = true, features = ["derive"] }
 strum_macros = { workspace = true, optional = true }
+tempfile = { workspace = true, optional = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, optional = true, features = ["rt", "time"] }
 tracing.workspace = true
@@ -48,11 +52,12 @@ uint.workspace = true
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt", "time"] }
+tempfile.workspace = true
 
 [features]
 default = ["strum"]
 float = []
-test-utils = ["dep:config"]
+test-utils = ["dep:config", "dep:tempfile"]
 agent = ["ethers", "strum"]
 strum = ["dep:strum"]
 ethers = ["dep:ethers-core", "dep:ethers-contract", "dep:ethers-providers", "dep:primitive-types"]
