+++ rust/hyperlane-core/src/db/mod.rs
@@ -1,0 +1,2 @@
+pub use rocks::*;
+mod rocks;
